requirements.txt
discord.py
python_dotenv
openai
flask